+++ examples/src/main/java/com/demcha/examples/GenerateAllExamples.java
@@ -96,6 +96,7 @@ import com.demcha.examples.templates.cv.v2.CvNordicCleanExample;
 import com.demcha.examples.templates.cv.v2.CvPanelExample;
 import com.demcha.examples.templates.cv.v2.CvSidebarPortraitExample;
 import com.demcha.examples.templates.cv.v2.CvTimelineMinimalExample;
+import com.demcha.examples.templates.invoice.ClassicInvoiceV2Example;
 import com.demcha.examples.templates.invoice.ConsultingInvoiceV2Example;
 import com.demcha.examples.templates.invoice.InvoiceCinematicFileExample;
 import com.demcha.examples.templates.invoice.ModernInvoiceV2Example;
@@ -163,6 +164,7 @@ public final class GenerateAllExamples {
         // Invoices
         System.out.println("Generated: " + InvoiceCinematicFileExample.generate());
         System.out.println("Generated: " + ModernInvoiceV2Example.generate());
+        System.out.println("Generated: " + ClassicInvoiceV2Example.generate());
         System.out.println("Generated: " + ConsultingInvoiceV2Example.generate());
 
         // Proposals